Updates `debug` package to address security alert
Addresses the [VDB-217665](https://security.snyk.io/vuln/SNYK-DEBIAN11-NODEDEBUG-3231713) vulnerability on versions matching <3.1.0-1.

Stayed within version 3.x instead of the latest v4. v4 does not install or functiion on lower versions of node listed in the workflow.

Verified that all of the interfaces used in the body-parser code are still supported. Also verified by running the tests and enabling the debug messages via:

```
DEBUG=* npm run test
```

Verified the tests pass locally on all the LTS versions but depending on the GH Action to run against all node versions supported to verify it continues to work for them.